requirements.txt
playwright==1.28.0
pytest==7.2.0
pytest-rerunfailures==11.0
allure-pytest==2.12.0
allure-python-commons==2.12.0